requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
fastapi==0.115.12
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pydantic==2.11.2
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.46.1
supabase==1.2.0
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
